requirements.txt
opencv-python
cupy-cuda11x
tifffile
scikit-image
nibabel
zarr
tqdm
pandas
openpyxl
seaborn
csbdeep
stardist
gputools
hist
opencv-python
dask[distributed]
dask-image
dask
porespy
microfilm
n2v
dask-cuda
pyvis